1.1.4.2 - 1.1.4.5 Combined Update
* `ImageBoxEx` example control now correctly allows you to move and
resize selection regions even when the control is zoomed
* If the `InterpolationMode` property is set to `Default`, the
`ImageBox` control will now choose an appropriate mode based on the
current zoom levels
* Added new `DrawBackground` virtual method. This allows you to override
how the `ImageBox` draws the image background without having to override
the entire `OnPaint` method.
* Added a new `TextPadding` property and corresponding `DrawLabel`
overload. When this property is set, rendered text from the `Text` and
`TextAlign` properties is drawn with the appropriate padding. If the
`TextBackColor` property is set to something not fully transparent, then
the background will be filled in the original render spot, excluding the
padding, allowing for labels with wider background borders.